Updated to JDA 5.2.1, removed text watching and message commands
honestly I didn't mean to remove message commands (I was only focused on removing message reading related code), but I will leave it removed for now because in FBSR it was very rarely used and no one knew it even existed as an option.  So this is now a way to simplify the code.  Maybe one day DCBA will be deprecated and FBSR will use JDA directly in the future.